+++ src/modules/generator.js
@@ -4,6 +4,8 @@ const KareeGeneratorMeta = require('../models/karee_generator_meta')
 const {__karee_helper, KareeProjectConfig} = require('../models/karee_config')
 const validator = require('../tools/validator')
 const projectConstants = require('../config/constants')
+const KareeProjectConstants = require('../config/constants')
+
 module.exports = class KareeGenerator{
 
     settings = new KareeGeneratorMeta()
@@ -26,7 +28,6 @@ module.exports = class KareeGenerator{
         }else if(this.settings.isScreen){
             this.generateScreen()
         }
-        // console.log(this.settings)
     }
 
     /**
@@ -48,72 +49,56 @@ module.exports = class KareeGenerator{
     generateController(){
         
         validator.validateGeneratedController(this.settings)
-        
+        const classNameToCambelCase = formatter.underscoreToCambel(this.settings.className)
         let template = io.readFile(io.projectFile(KareeProjectConfig.__template_controller), false) 
+
         template = template
             .toString()
             .replace('$appName', this.projectConfig.appName)
-            .replace('$className', formatter.underscoreToCambel(this.settings.className))
-            .replace('$className', formatter.underscoreToCambel(this.settings.className))
-            .replace('$className', formatter.underscoreToCambel(this.settings.className))
-            .replace('$className', formatter.underscoreToCambel(this.settings.className))
-            .replace('$className', formatter.underscoreToCambel(this.settings.className))
-        
+            .replaceAll('$className', classNameToCambelCase)
         /**
-         * On crée l'arborescence indiquée par l'option --path où générer le controller
+         * We create the whole path passed via --path option.
          */
+        const controllerPath = KareeProjectConstants.app_base_dir+(this.settings.path == '/' ? '' : this.settings.path)
+        io.createDir(controllerPath)
+        io.writeFile(template, `${controllerPath}${formatter.cambelToUnderscore(this.settings.className)}_controller.dart`)
 
-        io.createDir('lib/app/controllers/'+(this.settings.path == '/' ? '' : this.settings.path))
-        io.writeFile(template, `lib/app/controllers/${(this.settings.path == '/' ? '' : this.settings.path)}${formatter.cambelToUnderscore(this.settings.className)}_controller.dart`)
-
-        console.log(`\n\x1b[36m${this.settings.className}Controller\x1b[0m generated in lib/app/controllers/${(this.settings.path == '/' ? '' : this.settings.path)}${formatter.cambelToUnderscore(this.settings.className)}_controller.dart\n`)
+        console.log(`\n\x1b[36m${classNameToCambelCase}Controller\x1b[0m generated in lib/app/controllers/${(this.settings.path == '/' ? '' : this.settings.path)}${formatter.cambelToUnderscore(this.settings.className)}_controller.dart\n`)
 
     }
 
     generateScreen(){
-        // console.log(this.settings)
         validator.validateGeneratedScreen(this.settings)
         let template = ''
-        if(this.settings.isStatefull)
+        if(this.settings.isStatefull){
             template = io.readFile(io.projectFile(KareeProjectConfig.__template_stf_screen), false) 
-        else
-            template = io.readFile(io.projectFile(KareeProjectConfig.__template_stl_screen), false)
-        template = template
-            .toString()
-            .replace('$appName', this.projectConfig.appName)
-            .replace('$className', formatter.underscoreToCambel(this.settings.className))
-            .replace('$className', formatter.underscoreToCambel(this.settings.className))
-            .replace('$className', formatter.underscoreToCambel(this.settings.className))
-        if(this.settings.name === null || this.settings.name === undefined){
-            template = template 
-                .replace('@Screen("$name")', formatter.cambelToUnderscore(''))
-                .replace('import \'package:karee/annotations.dart\';\n', '')
-                .replace('$name', formatter.cambelToUnderscore(''))
-            // console.log(template)
         }else{
-            template = template
-                .replace('$name', formatter.cambelToUnderscore(this.settings.name))
-                .replace('$name', formatter.cambelToUnderscore(this.settings.name))
+            template = io.readFile(io.projectFile(KareeProjectConfig.__template_stl_screen), false)
         }
 
-        if(this.settings.isStatefull)
-            template = template
-                .replace('$className', formatter.underscoreToCambel(this.settings.className))
-                .replace('$className', formatter.underscoreToCambel(this.settings.className))
-                .replace('$className', formatter.underscoreToCambel(this.settings.className))
-        /**
-         * On Crée l'arborescence indiqué par l'option --path où ajouter l'écran
-         */
-        io.createDir('lib/app/screens/'+(this.settings.path == '/' ? '' : this.settings.path))
-        io.writeFile(template, `${'lib/app/screens/'+(this.settings.path == '/' ? '' : this.settings.path)}${formatter.cambelToUnderscore(this.settings.className)}_screen.dart`)
+        const classNameToCambelCase = formatter.underscoreToCambel(this.settings.className)
 
-        /**
-         * On crée un composant test
-         */
-        io.createDir('lib/app/screens/'+this.settings.path+'/components')
-        io.copy(io.projectFile(KareeProjectConfig.__template_components), 'lib/app/screens/'+(this.settings.path == '/' ? '' : this.settings.path)+'components/congrat_card.dart')
+        template = template
+            .toString()
+            .replaceAll('$className', classNameToCambelCase)
 
-        console.log(`\n\x1b[36m${formatter.underscoreToCambel(this.settings.className)}Screen\x1b[0m generated in lib/app/screens/${(this.settings.path == '/' ? '' : this.settings.path)}${formatter.cambelToUnderscore(this.settings.className)}_screen.dart\n`)
+        if(this.settings.name === null || this.settings.name === undefined){
+            template = template 
+                .replace('@Screen(\'$ScreenName\')\n', '')
+                .replace('import \'package:karee/annotations.dart\';\n', '')
+        }else{
+            template = template.replace('$ScreenName', formatter.cambelToUnderscore(this.settings.name))
+        }
+
+        const screenPath = KareeProjectConstants.app_base_dir+(this.settings.path == '/' ? '' : this.settings.path)
+        const classNameToUnderscoreCase = formatter.cambelToUnderscore(this.settings.className)
+        /**
+         * Here we create the three passed as --path option where the screen should be created.
+         */
+        io.createDir(screenPath)
+        io.writeFile(template, `${screenPath}${classNameToUnderscoreCase}_screen.dart`)
+
+        console.log(`\n\x1b[36m${classNameToCambelCase}Screen\x1b[0m generated in ${screenPath}${classNameToUnderscoreCase}_screen.dart\n`)
     }
 
 }