+++ src/modules/run_build.js
@@ -39,13 +39,26 @@ class KareeSourceGen{
         let generator = new KareeGenerator();
         generator.generateResources()
 
-        cmd.run(`flutter clean && flutter pub get && flutter packages pub run build_runner watch --delete-conflicting-outputs`)
-           .on('close', (code, signal) => {
-               spinner.stop(false)
-               console.log('\n')
-               callback?.call()
-           })
-           .on('message', (msg) => console.log(msg))
+//        spinner = new Spinner()
+//        spinner.setSpinnerString('⠋⠙⠹⠸⠼⠴⠦⠧⠇⠏')
+        setTimeout(() => {
+            spinner.setSpinnerTitle(`\x1b[32m\x1b[1mKaree is running flutter clean in \x1b[33m${this.projectConfig.appName}\x1b[39m\x1b[22m\x1b[0m`)
+            spinner.start()
+    
+            cmd.run(`flutter clean`)
+               .on('close', (code, signal) => {
+                    spinner.setSpinnerTitle(`\x1b[32m\x1b[1mKaree is running flutter pub get in \x1b[33m${this.projectConfig.appName}\x1b[39m\x1b[22m\x1b[0m`)
+                    spinner.start()
+                    cmd.run(`flutter pub get && dart run build_runner watch --delete-conflicting-outputs`)
+                        .on('close', (code, signal) => {
+                            spinner.stop(false)
+                            console.log('\n')
+                            callback?.call()
+                        })
+                })
+               .on('error', (_) => console.log(_))
+               .on('message', (msg) => console.log(msg))
+            }, 1000)
     }
 }